langchain_learning/requirements.txt
langchain
langchain-openai
langchain-google-genai
langchain-community
python-dotenv
